+++ src/textual/app.py
@@ -156,7 +156,23 @@ class ScreenError(Exception):
 
 
 class ScreenStackError(ScreenError):
-    """Raised when attempting to pop the last screen from the stack."""
+    """Raised when trying to manipulate the screen stack incorrectly."""
+
+
+class ModeError(Exception):
+    """Base class for exceptions related to modes."""
+
+
+class InvalidModeError(ModeError):
+    """Raised if there is an issue with a mode name."""
+
+
+class UnknownModeError(ModeError):
+    """Raised when attempting to use a mode that is not known."""
+
+
+class ActiveModeError(ModeError):
+    """Raised when attempting to remove the currently active mode."""
 
 
 class ModeError(Exception):
@@ -652,14 +668,7 @@ class App(Generic[ReturnType], DOMNode):
         """
         self.set_class(dark, "-dark-mode")
         self.set_class(not dark, "-light-mode")
-        try:
-            self.refresh_css()
-        except ScreenStackError:
-            # It's possible that `dark` can be set before we have a default
-            # screen, in an app's `on_load`, for example. So let's eat the
-            # ScreenStackError -- the above styles will be handled once the
-            # screen is spun up anyway.
-            pass
+        self.call_later(self.refresh_css)
 
     def get_driver_class(self) -> Type[Driver]:
         """Get a driver class for this platform.
@@ -1550,7 +1559,7 @@ class App(Generic[ReturnType], DOMNode):
 
         Args:
             screen: A Screen instance or the name of an installed screen.
-            callback: An optional callback function that is called if the screen is dismissed with a result.
+            callback: An optional callback function that will be called if the screen is [dismissed][textual.screen.Screen.dismiss] with a result.
 
         Returns:
             An optional awaitable that awaits the mounting of the screen and its children.
@@ -1766,8 +1775,22 @@ class App(Generic[ReturnType], DOMNode):
 
     def _print_error_renderables(self) -> None:
         """Print and clear exit renderables."""
-        for renderable in self._exit_renderables:
-            self.error_console.print(renderable)
+        error_count = len(self._exit_renderables)
+        if "debug" in self.features:
+            for renderable in self._exit_renderables:
+                self.error_console.print(renderable)
+            if error_count > 1:
+                self.error_console.print(
+                    f"\n[b]NOTE:[/b] {error_count} errors show above.", markup=True
+                )
+        elif self._exit_renderables:
+            self.error_console.print(self._exit_renderables[0])
+            if error_count > 1:
+                self.error_console.print(
+                    f"\n[b]NOTE:[/b] 1 of {error_count} errors show. Run with [b]--dev[/] to see all errors.",
+                    markup=True,
+                )
+
         self._exit_renderables.clear()
 
     async def _process_messages(
@@ -2269,6 +2292,7 @@ class App(Generic[ReturnType], DOMNode):
             screen = Screen(id=f"_default")
             self._register(self, screen)
             self._screen_stacks[self._current_mode].append(screen)
+            screen.post_message(events.ScreenResume())
             await super().on_event(event)
 
         elif isinstance(event, events.InputEvent) and not event.is_forwarded: